add support for null event and couple of new usages
nullevent is an event that can happen in the steam but we want to skip
it.

added two sample events like that: INTVAR_EVENT and GTID_LOG_EVENT -
those are events that happen when we do GTID enabled replication, but
are of no use (at least not for me at this time) in this packages
aspect. It might be when we support moving to a different master
seamlessly.