Partial fix for dropped columns blowing up replication when replaying binlog with past events
Mitigates following errror:
  File "/usr/local/lib/python2.7/dist-packages/pymysqlreplication/binlogstream.py", line 262, in fetchone
    self.__freeze_schema)
  File "/usr/local/lib/python2.7/dist-packages/pymysqlreplication/packet.py", line 98, in __init__
    freeze_schema = freeze_schema)
  File "/usr/local/lib/python2.7/dist-packages/pymysqlreplication/row_event.py", line 550, in __init__
    column_schema = self.column_schemas[i]
IndexError: list index out of range

This is just a superficial patch. It does not address deeper issue at hand, which is that pymysqlreplication
has no good way of handling RowEvents containing columns which have been dropped prior to pymsqlreplication
start. The problem is always relying on current version of information_schema even though dealing with events
that might be several versions in the past.