Implement Heartbeat
I ended up debugging a slow startup with pymysqlreplication,
the stream "stucked" for a few seconds before seeing first events coming in.
For that purpose I ended up implementing HeartbeatLogEvent and
documenting observed mysql behavior.

I also implemented the heartbeat feature, which is required for the mysqld
to filter events in gtid mode.

Signed-off-by: Arthur Gautier <baloo@gandi.net>