Clear table_map if RotateEvent has timestamp of 0 (#197)
* Clear table_map if RotateEvent has timestamp of 0

MySQL generates two RotateEvents when flushing binary logs.  The first
is added to the end of the old binary log.  This RotateEvent has a valid
timestamp.  The second is added to the beginning of the new binary log.
This RotateEvent has a timestamp of 0.

Under some circumstances, e.g., the MySQL server crashes, the first
RotateEvent will not be written to the binary log.  (See the MySQL
Internals Manual 20.9.)

Before this change, when instantiating BinLogStreamReader with a
positive skip_to_timestamp argument, the RotateEvents with a timestamp
of 0 were ignored.  This could cause silent corruption of the table_map.

This change ensures that RotateEvents always reset the table_map,
log_pos and log_file attributes.

* Update documentation on running tests

* Run the test_no_trailing_rotate_event with sudo

This change alters the Travis CI configuration to run the
test_no_trailing_rotate_event with sudo and exclude it from the normal
test run.

* Fix sudo not finding the nosetests command