fix bug
timing of event creation and the timing of querying the table structure are different,
ex) drop table but replication did not reach that point 
f self.column_schemas doesn't yield any results, the code will exit the if statement. However, since there is more column information in the remaining packets, if the column.count is not 0, we still need to read the packets.